fixes #45: send length, image, upgrade if offset is 0
This fix causes the upload routine to conform to
the SMP server's expectations in the case where
the initial packet contains no image data.  This
should allow for several packets to be sent with
offset=0.  When offset=0 and some image data is
present, this commit ensures that the length,
image, and upgrade fields are still included, even
if it has already been sent.