fix: ble: Race condition in receive for notifications
Fix a race condition relating to BLE notifications. If the notification arrives before the `receive()` method is called, the notification will be lost and the transfer will time out. This happens because `self._notify_condition.notify()` does nothing without a pending `self._notify_condition.wait()`.

Only wait for notifications if the buffer does not contain enough data for the header.