fix: Fix single byte being parsed as serial.
Single bytes cannot be flushed to serial if they may be starts of SMP packets. Technically this introduces an edge case where 0x04 or 0x06 is the very last character of a valid serial data stream, and will then not get flushed properly, but the chances of that are very small, and there aren't any good workarounds to this.